Test,exe/: Build system and sample JR-200 "Hello" program
The build system using b8tool is set up per the description in the README.
Currently it builds just a stand-alone program, exe/jr200/hello.a68. That
program makes use of code in 8bitdev, but since we're using a whole program
build we need no special support for building code from 8bitdev; we just
put `8bitdev/src/…` at the front of the paths we `include`.

We do not yet support unit tests in the top-level module (typically under
src/) ; that requires some further changes and will come in the next
commit. We do run the b8tool unit tests. We do not build and test unit-test
versions of the code in submodules (8bitdev); that requires some extensive
thought and work. This is further documented in the comments in `Test`.