8bitdev/ → cjs8bitdev; add src/ unit tests
Because Python doesn't allow module names to start with a digit, we rename
the `8bitdev` submodule to `cjs8bitdev`.

This allows us to add an src/conftest.py that imports the assembly-language
unit test setup code from the cjs8bitdev.src.conftest module. (That module
needs to be moved to b8tool.)

That in turn lets us add a very simple src/sample.{a68,pt} that shows
the assembly-language unit testing to be working for this project.
Run `./Test -vv src/` to see the test discovery and execution for these.